refactor: introduce game core
