+++ roguelike_pygame.py
@@ -1,12 +1,12 @@
 import pygame
 import sys
 from roguelike_ai import generate_dungeon, Player, GameState, handle_command, ROOM_TYPES, ENEMIES, LOOT_TABLE, openai, api_call_counter, get_api_call_counter
+from core import config
+from core.game import Game
 import random
-import copy
 import os
 import time  # Import time module for timeout logic
 import json
-from collections import deque
 
 # Ensure OpenAI API key is set
 OPENAI_API_KEY = os.getenv("OPENAI_API_KEY")
@@ -20,10 +20,10 @@ openai.api_key = OPENAI_API_KEY
 
 # --- Pygame Setup ---
 pygame.init()
-CELL_SIZE = 120  # Even larger
-GRID_W, GRID_H = 6, 6
-STATUS_BAR_HEIGHT = 40
-WIDTH, HEIGHT = CELL_SIZE * GRID_W + 80 + 320, CELL_SIZE * GRID_H + 400 + STATUS_BAR_HEIGHT  # Add status bar height
+CELL_SIZE = config.CELL_SIZE
+GRID_W, GRID_H = config.GRID_W, config.GRID_H
+STATUS_BAR_HEIGHT = config.STATUS_BAR_HEIGHT
+WIDTH, HEIGHT = config.WIDTH, config.HEIGHT
 screen = pygame.display.set_mode((WIDTH, HEIGHT))
 pygame.display.set_caption("DungeonGPT Roguelike")
 font = pygame.font.SysFont(None, 36)  # Larger font
@@ -103,20 +103,23 @@ seed = 1337
 rooms = generate_grid_dungeon(seed, GRID_W, GRID_H)
 player = Player(location="room_0")
 state = GameState(seed=seed, rooms=rooms, player=player)
+game = Game(state=state, enemy_sprites=enemy_sprites)
+# Alias for convenience in the existing code
+state = game.state
 
 # --- Colors ---
-BLACK = (20, 20, 30)
-WHITE = (240, 240, 240)
-GRAY = (60, 60, 80)
-BLUE = (80, 180, 255)
-GREEN = (80, 220, 120)
-RED = (255, 80, 80)
-YELLOW = (255, 220, 80)
-ROOM_BORDER = (180, 180, 220)
-CONNECTION = (60, 60, 120)
+BLACK = config.BLACK
+WHITE = config.WHITE
+GRAY = config.GRAY
+BLUE = config.BLUE
+GREEN = config.GREEN
+RED = config.RED
+YELLOW = config.YELLOW
+ROOM_BORDER = config.ROOM_BORDER
+CONNECTION = config.CONNECTION
 
 # --- Stats Panel Setup ---
-STATS_PANEL_WIDTH = 320
+STATS_PANEL_WIDTH = config.STATS_PANEL_WIDTH
 stats_panel = pygame.Rect(CELL_SIZE * GRID_W + 80, 0, STATS_PANEL_WIDTH, HEIGHT)
 
 # --- Text Input State ---
@@ -192,95 +195,7 @@ backgrounds = load_backgrounds()
 selected_bg = random.choice(backgrounds) if backgrounds else None
 
 # --- Enemy Movement and Spawning Helpers ---
-def move_non_boss_enemies(state):
-    grid_w, grid_h = GRID_W, GRID_H
-    new_rooms = copy.deepcopy(state.rooms)
-    coord_to_room = {room['coords']: rid for rid, room in state.rooms.items()}
-    boss_coords = {room['coords'] for room in state.rooms.values() if room['type'] == 'boss_room'}
-    occupied = {
-        room['coords']
-        for room in state.rooms.values()
-        for e in room['enemies']
-        if e['name'] != 'dungeon_boss'
-    }
-    player_coord = state.rooms[state.player.location]['coords']
-    directions = [(-1, 0), (1, 0), (0, -1), (0, 1)]
-
-    def bfs(start, goal, blocked):
-        if start == goal:
-            return []
-        queue = deque([start])
-        came_from = {start: None}
-        while queue:
-            current = queue.popleft()
-            if current == goal:
-                break
-            for dx, dy in directions:
-                nx, ny = current[0] + dx, current[1] + dy
-                if 0 <= nx < grid_w and 0 <= ny < grid_h:
-                    nxt = (nx, ny)
-                    if nxt not in came_from and nxt not in blocked:
-                        queue.append(nxt)
-                        came_from[nxt] = current
-        if goal not in came_from:
-            return None
-        path = []
-        cur = goal
-        while cur != start:
-            path.append(cur)
-            cur = came_from[cur]
-        path.reverse()
-        return path
-
-    for room_id, room in state.rooms.items():
-        for enemy in room['enemies'][:]:
-            if enemy['name'] == 'dungeon_boss':
-                continue  # Boss does not move
-            start = room['coords']
-            occupied.discard(start)
-            blocked = occupied | boss_coords
-            path = bfs(start, player_coord, blocked)
-            if path:
-                next_step = path[0]
-                dest_id = coord_to_room[next_step]
-                if not new_rooms[dest_id]['enemies']:
-                    new_rooms[room_id]['enemies'].remove(enemy)
-                    new_rooms[dest_id]['enemies'].append(enemy)
-                    occupied.add(next_step)
-                else:
-                    occupied.add(start)
-            else:
-                occupied.add(start)
-    state.rooms = new_rooms
-    return state
-
-def count_non_boss_enemies(state):
-    count = 0
-    for room in state.rooms.values():
-        for enemy in room['enemies']:
-            if enemy['name'] != 'dungeon_boss':
-                count += 1
-    return count
-
-def spawn_enemy(state):
-    # Spawn a new non-boss enemy in a random room not occupied by player or boss
-    possible_rooms = [rid for rid, room in state.rooms.items()
-                     if room['type'] != 'boss_room' and rid != state.player.location]
-    if not possible_rooms:
-        return state
-    room_id = random.choice(possible_rooms)
-    name = random.choice([k for k in ENEMIES if k != 'dungeon_boss'])
-    enemy = {**ENEMIES[name], 'name': name}
-    # Assign a random sprite for this enemy
-    sprite_keys = [k for k in enemy_sprites if name in k]
-    if sprite_keys:
-        enemy['sprite'] = random.choice(sprite_keys)
-    else:
-        enemy['sprite'] = None
-    new_rooms = copy.deepcopy(state.rooms)
-    new_rooms[room_id]['enemies'].append(enemy)
-    state.rooms = new_rooms
-    return state
+# These helpers have been moved into the :class:`core.game.Game` class.
 
 # --- Hero Selection Dialog ---
 # Mapping of hero names to stat blocks and abilities
@@ -343,20 +258,7 @@ def animate_d20_roll(screen, font, roll_result):
     pygame.display.flip()
     pygame.time.delay(1000)
 
-# --- Integrate D20 Roll into Attack Logic ---
-def attack_enemy(state, enemy):
-    """Roll a d20 and apply damage to the given enemy.
-
-    Returns a tuple of (message, roll_result) so the caller can display the
-    roll in the output area and forward the result to the AI for consistent
-    narration.
-    """
-    roll_result = roll_d20()
-    animate_d20_roll(screen, font, roll_result)
-    damage = roll_result + state.player.str  # Example: Add player's strength to roll
-    enemy["hp"] -= damage
-    msg = f"You rolled a {roll_result} and dealt {damage} damage to {enemy['name']}!"
-    return msg, roll_result
+# Attack logic handled by Game.attack_enemy; UI performs roll and animation.
 
 # --- Main Loop ---
 running = True
@@ -624,7 +526,9 @@ while running:
                 room = state.rooms.get(state.player.location, {})
                 target_enemy = next((e for e in room.get('enemies', []) if e.get('hp', 0) > 0), None)
                 if target_enemy:
-                    atk_msg, roll_result_to_send = attack_enemy(state, target_enemy)
+                    roll_result_to_send = roll_d20()
+                    animate_d20_roll(screen, font, roll_result_to_send)
+                    atk_msg = game.attack_enemy(target_enemy, roll_result_to_send)
                     for line in wrap_text(atk_msg, output_font, output_area.width-32):
                         output_lines.append(line)
                     if target_enemy["hp"] <= 0:
@@ -685,9 +589,9 @@ while running:
         input_text = ""  # Clear input after response
         # Only move/spawn enemies if not in battle dialog
         if not battle_options:
-            state = move_non_boss_enemies(state)
-            while count_non_boss_enemies(state) < 2:
-                state = spawn_enemy(state)
+            game.move_non_boss_enemies()
+            while game.count_non_boss_enemies() < 2:
+                game.spawn_enemy()
         # Flash if entered a room with enemies
         new_room = state.rooms.get(state.player.location, {})
         if state.player.location != prev_location and new_room.get("enemies"):
@@ -738,7 +642,8 @@ while running:
                                     seed = random.randint(1, 999999)
                                     rooms = generate_grid_dungeon(seed, GRID_W, GRID_H)
                                     player = Player(location="room_0")
-                                    state = GameState(seed=seed, rooms=rooms, player=player)
+                                    game.state = GameState(seed=seed, rooms=rooms, player=player)
+                                    state = game.state
                                     selected_hero = select_hero()
                                     player_icon = hero_sprites[selected_hero]
                                     game_over = False
@@ -768,7 +673,9 @@ while running:
                 room = state.rooms.get(state.player.location, {})
                 target_enemy = next((e for e in room.get('enemies', []) if e.get('hp', 0) > 0), None)
                 if target_enemy:
-                    atk_msg, roll_result_to_send = attack_enemy(state, target_enemy)
+                    roll_result_to_send = roll_d20()
+                    animate_d20_roll(screen, font, roll_result_to_send)
+                    atk_msg = game.attack_enemy(target_enemy, roll_result_to_send)
                     for line in wrap_text(atk_msg, output_font, output_area.width-32):
                         output_lines.append(line)
                     if target_enemy["hp"] <= 0:
@@ -828,9 +735,9 @@ while running:
         output_scroll = 0
         input_text = ""
         if not battle_options:
-            state = move_non_boss_enemies(state)
-            while count_non_boss_enemies(state) < 2:
-                state = spawn_enemy(state)
+            game.move_non_boss_enemies()
+            while game.count_non_boss_enemies() < 2:
+                game.spawn_enemy()
         new_room = state.rooms.get(state.player.location, {})
         if state.player.location != prev_location and new_room.get("enemies"):
             flash_timer = pygame.time.get_ticks() + flash_duration
@@ -879,7 +786,8 @@ while running:
                                     seed = random.randint(1, 999999)
                                     rooms = generate_grid_dungeon(seed, GRID_W, GRID_H)
                                     player = Player(location="room_0")
-                                    state = GameState(seed=seed, rooms=rooms, player=player)
+                                    game.state = GameState(seed=seed, rooms=rooms, player=player)
+                                    state = game.state
                                     selected_hero = select_hero()
                                     player_icon = hero_sprites[selected_hero]
                                     game_over = False
